+++ src/main/java/com/meetplus/batch/job/BeforeEventStartJob.java
@@ -1,13 +1,11 @@
-package com.meetplus.batch.schedule;
+package com.meetplus.batch.job;
 
 import com.meetplus.batch.kafka.KafkaProducerCluster;
 import com.meetplus.batch.kafka.Topics;
 import com.meetplus.batch.kafka.dto.EventPreviewDto;
-import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
-import org.springframework.batch.core.configuration.annotation.EnableBatchProcessing;
 import org.springframework.batch.core.job.builder.JobBuilder;
 import org.springframework.batch.core.repository.JobRepository;
 import org.springframework.batch.core.step.builder.StepBuilder;
